pyproject.toml
[build-system]
requires = ["uv_build>=0.10.5,<0.11.0"]
build-backend = "uv_build"

[project]
name = "amelis-smime-decrypt"
version = "0.2.0"
description = "A library and CLI tool to fetch and decrypt S/MIME encrypted emails and extract PDF attachments."
authors = [
    { name = "Christian Gonzalez", email = "office@nerdocs.at" }
]
license = { text = "GPL-3.0-or-later" }
readme = "README.md"
requires-python = ">=3.8"
keywords = ["email", "smime", "decryption", "imap", "attachments"]

dependencies = [
    "imapclient",
    "python-dotenv",
    "cryptography",
    "endesive",
]

[project.optional-dependencies]
dev = [
    "black",
#    "pytest",
#    "pytest-cov",
]

[project.scripts]
amelis-smime-decrypt = "amelis_smime_decrypt.cli:main"

[project.urls]
Homepage = "https://github.com/nerdocs/amelis-smime-decrypt"
Repository = "https://github.com/nerdocs/amelis-smime-decrypt"
Issues = "https://github.com/nerdocs/amelis-smime-decrypt/issues"
